consider power state before acting
This patch adds state handling and status output relating to it in the
various actions. For example, we only want to power off a vm if it's
currently on and we always want to ensure a vm is on before trying to
provision, etc.

This makes the halt action compatible with other actions. Previously,
once you halted, you weren't able to get the vm up and running (through
vagrant) again.

Signed-off-by: Anton Lofgren <alofgren@op5.com>