clone: make it possible to clone from a VM instead of a template
This is useful when you don't have DRS enabled, and thus are prevented
from using templates as the cloning source.

I've implemented this to have as little impact as possible on the
existing implementation by using a new configuration option
"clone_from_vm". If this option is true, vagrant-vsphere (perphaps
somewhat unintuitively?) reuses the template_name as the name of the
VM to clone from and disables any error for resource pool and compute
resource. In short, the only thing we need to do in order to clone from
a VM is to not specify a pool for the VirtualMachineRelocateSpec.