Add specs for reload action.
Add 'reload' tests mimicking the tests for 'halt' and 'up', with the
difference that 'reload' does *not* create the VM if it does not exist.

Also fix the class name in `lib/vSphere/action/message_not_created.rb`,
which was needed for one of the new specs to pass (and likely caused
other problems as well).