Add support for setting guestinfo variables
This patch adds an `extra_config` section to the provider configuration which
can be set to a hash. This hash is used to populate the `extraConfig` section
of the `VirtualMachineConfigSpec` used to clone new VMs. The keys of the hash
must start with `guestinfo.` and the values are available to guest VMs with
VMware tools installed via the `vmtoolsd` command.