Add ‘real_nic_ip’ option to allow users to enable interface filtering behavior when obtaining the ssh IP address
This commit adds an option that forces get_ssh_ip to filter the vm.guest.net entries for a network adapter that ‘exists’ on the VM, as vsphere does not always return the correct primary ipAddress if new interfaces have been defined at provision time or later.   We encountered the situation at NSIDC when provisioning VMs which were running docker/docker-compose and had docker-defined internal interfaces that vsphere picked up and assumed were the primary IP.  

This will not address situations where users have defined a VM with multiple ‘legitimate’ interfaces, and throws an error if that situation is encountered.