Moved helpers into driver and changing actions to use driver. Connection is now no longer exposed to actions.
A driver is created per a machine, and it manages the connection itself, so it is now safe between multiple machine deployments.

Removed calling @machine.action, so we don't create a lock on the actions anymore, as they are now safe to call in parallel.

Moved static strings into locales/en.yml