Preserve backtraces when re-raising errors
This commit updates locations where StandardError instances are rescued
and re-raised as VSphereError to preserve the backtrace info of the
original errors along with the message. This change makes debugging
easier as the backtrace leads to the location where the error originated
instead of the rescue block where it was re-raised.